linter/lintmain: re-design enable/disable API
-enable includes checkers, -disable excludes them.
By default, -enable is filled by all checkers that
don't have performance, opinionated or experimental tag.

There is also -enableAll flag that is identical to
manually enumerating every checker in -enable.
If -enableAll is used, -enable is effectively ignored.

Both -enable and -disable can use tags as named groups
of checkers. The syntax is "#"+tag name.
For example, including only diagnostics can be done with
	-enable='#diagnostic'.
Excluding opinionated and experimental can be achieved with
	-disable='#opinionated','#experimental'

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>